style: update some export types of TaggedAIAsset
