+++ spx-gui/src/components/asset/library/AssetLibrary.vue
@@ -92,7 +92,8 @@ import { asset2Backdrop, asset2Sound, asset2Sprite, type AssetModel } from '@/mo
 import { NIcon, NSpin } from 'naive-ui'
 import { BulbOutlined } from '@vicons/antd'
 import { useSearchCtx, useSearchResultCtx } from './SearchContextProvider.vue'
-import AssetList, { isAiAsset, type AssetOrAIAsset, type TaggedAIAssetData } from './AssetList.vue'
+import AssetList from './AssetList.vue'
+import { isAiAsset, type AssetOrAIAsset, type TaggedAIAssetData } from '@/apis/aigc'
 import LibraryMenu from './LibraryMenu.vue'
 import LibraryTree from './LibraryTree.vue'
 import LibraryTab from './LibraryTab.vue'
